+++ crates/api-desc/src/store/fragment.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Google LLC
+// Copyright 2023 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,42 +14,24 @@
 
 use crate::*;
 
-mod fragment;
-
 pub(crate) fn new() -> Item {
     let docs = docs! {
-        /// Persistent storage operations.
+        /// Support for fragmented entries.
     };
-    let name = "store".into();
+    let name = "fragment".into();
     let items = vec![
-        item! {
-            /// Describes errors interacting with the store.
-            enum Error {
-                /// A function pre-condition was broken.
-                InvalidArgument = 0,
-
-                /// The store is full.
-                NoCapacity = 1,
-
-                /// The store reached its end of life.
-                NoLifetime = 2,
-
-                /// An operation to the underlying storage failed.
-                StorageError = 3,
-
-                /// The underlying storage doesn't match the store invariant.
-                InvalidStorage = 4,
-            }
-        },
         item! {
             /// Inserts an entry in the store.
             ///
-            /// If an entry for that key was already present, it is overwritten.
-            fn insert "si" {
-                /// Key of the entry.
+            /// The entry will be fragmented over multiple keys within the provided range as needed.
+            ///
+            /// If an entry for that range of keys was already present, it is overwritten.
+            fn insert "sfi" {
+                /// Range of keys where to insert the fragments.
                 ///
-                /// This must be smaller than 4096.
-                key: usize,
+                /// This is a pair of u16: the lowest u16 is the first key of the range and the
+                /// highest u16 is one past the last key of the range.
+                keys: u32,
 
                 /// Value of the entry.
                 ptr: *const u8,
@@ -64,10 +46,12 @@ pub(crate) fn new() -> Item {
         item! {
             /// Removes an entry from the store.
             ///
+            /// All fragments from the range of keys will be deleted.
+            ///
             /// This is not an error if no entry is present. This is simply a no-op in that case.
-            fn remove "sr" {
-                /// Key of the entry.
-                key: usize,
+            fn remove "sfr" {
+                /// Range of keys to remove.
+                keys: u32,
             } -> {
                 /// Zero for success. Otherwise complement of error number.
                 res: isize,
@@ -75,38 +59,25 @@ pub(crate) fn new() -> Item {
         },
         item! {
             /// Finds an entry in the store, if any.
-            fn find "sf" {
-                /// Key of the entry to find.
-                key: usize,
+            ///
+            /// The entry may be fragmented withen the provided range.
+            fn find "sff" {
+                /// Range of keys to concatenate as an entry.
+                keys: u32,
 
                 /// Where to write the value of the entry, if found.
                 ///
                 /// The (inner) pointer will be allocated by the callee and must be freed by the
                 /// caller. It is thus owned by the caller when the function returns.
-                #[cfg(not(feature = "multivalue"))]
                 ptr: *mut *mut u8,
 
                 /// Where to write the length of the value, if found.
-                #[cfg(not(feature = "multivalue"))]
                 len: *mut usize,
             } -> {
-                /// Value of the entry if found. Null if not found.
-                ///
-                /// The pointer is allocated by the callee and must be freed by the caller. It is
-                /// thus owned by the caller when the function returns.
-                #[cfg(feature = "multivalue")]
-                ptr: *mut u8,
-
-                /// Length of the value if non-negative. Otherwise complement of error number.
-                #[cfg(feature = "multivalue")]
-                len: isize,
-
                 /// One if found. Zero if not found. Otherwise complement of error number.
-                #[cfg(not(feature = "multivalue"))]
                 res: isize,
             }
         },
-        fragment::new(),
     ];
     Item::Mod(Mod { docs, name, items })
 }